Added the ability to add nicknames to classes
Added the ability to add nicknames to classes. However, since the database has no way to store nicknames right now, the nicknames aren't stored and are temporary. If the user navigates away from the screen, the name dissapaears and will need to be added back.
Also removed the delete button, as users shouldn't be able to delete classes with only the press of a button